refactor: extract URL parameter related logic to defaultGlobalService
