+++ packages/canvas/DesignCanvas/src/DesignCanvas.vue
@@ -229,7 +229,7 @@ export default {
       return key
     }
     const postUrlChanged = () => {
-      usePage().postLocationHistoryChanged(
+      getMetaApi(META_SERVICE.GlobalService).postLocationHistoryChanged(
         Object.fromEntries(
           Array.from(new URLSearchParams(window.location.search)).map(([key, value]) => [replaceKey(key), value])
         )
@@ -246,24 +246,6 @@ export default {
         subscriber: 'canvas_design_canvas'
       })
     })
-
-    function updatePreviewId(previewId, replace = false) {
-      const url = new URL(window.location.href)
-      if (previewId) {
-        if (previewId === url.searchParams.get('previewid')) {
-          return
-        }
-        url.searchParams.set('previewid', previewId)
-      } else {
-        url.searchParams.delete('previewid')
-      }
-      if (replace) {
-        window.history.replaceState({}, '', url)
-      } else {
-        window.history.pushState({}, '', url)
-      }
-      usePage().postLocationHistoryChanged({ previewId })
-    }
 
     return {
       removeNode,
@@ -282,7 +264,7 @@ export default {
         getPageAncestors: usePage().getAncestors,
         getBaseInfo: () => getMetaApi(META_SERVICE.GlobalService).getBaseInfo(),
         addHistoryDataChangedCallback,
-        updatePreviewId,
+        updatePreviewId: getMetaApi(META_SERVICE.GlobalService).updatePreviewId,
         ast,
         getBlockByName: useMaterial().getBlockByName,
         useModal,
