+++ packages/plugins/block/src/Main.vue
@@ -133,7 +133,15 @@ import {
 } from '@opentiny/vue'
 import { IconSearch } from '@opentiny/vue-icon'
 import { PluginPanel, PluginBlockList, SvgButton } from '@opentiny/tiny-engine-common'
-import { useBlock, useModal, useLayout, useCanvas, useHelp } from '@opentiny/tiny-engine-meta-register'
+import {
+  useBlock,
+  useModal,
+  useLayout,
+  useCanvas,
+  useHelp,
+  getMetaApi,
+  META_SERVICE
+} from '@opentiny/tiny-engine-meta-register'
 import { constants } from '@opentiny/tiny-engine-utils'
 import BlockSetting, { openPanel, closePanel } from './BlockSetting.vue'
 import BlockGroupArrange from './BlockGroupArrange.vue'
@@ -303,11 +311,7 @@ export default {
         useBlock().initBlock(block, {}, isEdit)
         useLayout().closePlugin()
         closePanel()
-        const url = new URL(window.location)
-        url.searchParams.delete('pageid')
-        url.searchParams.set('blockid', block.id)
-        window.history.pushState({}, '', url)
-        useBlock().postLocationHistoryChanged({ blockId: block.id })
+        getMetaApi(META_SERVICE.GlobalService).updateBlockId(block.id)
       } else {
         confirm({
           message: '当前画布内容尚未保存，是否要继续切换?',
